remove repo config
