+++ backend/infrahub/git/utils.py
@@ -3,7 +3,6 @@ from collections import defaultdict
 from typing import TYPE_CHECKING, Any
 
 from git import Repo
-from git import exc as git_exception
 from infrahub_sdk import InfrahubClient
 from infrahub_sdk.node import RelationshipManager
 from infrahub_sdk.protocols import CoreArtifactDefinition, CoreCheckDefinition, CoreGroup
@@ -183,19 +182,11 @@ def branch_name_in_import_sync_branches(branch_short_name: str) -> bool:
 
 def get_git_user_config(repo: Repo) -> tuple[str, str]:
     """
-    Checks for the git configuration of the user on the repo level
-    and if not found checks on the global level
+    Checks for the git configuration of the user on the global level
     Returns the git (username, email,)
     """
-    try:
-        with repo.config_reader() as git_config:
-            return (
-                str(git_config.get_value("user", "name")),
-                str(git_config.get_value("user", "email")),
-            )
-    except git_exception.GitCommandError:
-        with repo.config_reader(config_level="global") as git_config:
-            return (
-                str(git_config.get_value("user", "name")),
-                str(git_config.get_value("user", "email")),
-            )
+    with repo.config_reader(config_level="global") as git_config:
+        return (
+            str(git_config.get_value("user", "name")),
+            str(git_config.get_value("user", "email")),
+        )
