+++ backend/tests/conftest.py
@@ -1633,3 +1633,14 @@ async def schema_conversion_unidirectional_relationships(db: InfrahubDatabase, n
     }
 
     return schema
+
+
+@pytest.fixture
+def git_user_config():
+    initial_user_name = config.SETTINGS.git.user_name
+    initial_user_email = config.SETTINGS.git.user_email
+    config.SETTINGS.git.user_email = "test@email.com"
+    config.SETTINGS.git.user_name = "Test User"
+    yield
+    config.SETTINGS.git.user_email = initial_user_email
+    config.SETTINGS.git.user_name = initial_user_name
