+++ backend/tests/unit/git/conftest.py
@@ -1373,8 +1373,8 @@ async def mock_create_branch_git_repo_03(db: InfrahubDatabase, default_branch: B
     await create_branch(branch_name="branch01", db=db)
 
 
-@pytest.fixture
 def git_use_explicit_merge_commit_config():
+    initial_use_explicit_merge_commit = config.SETTINGS.git.use_explicit_merge_commit
     config.SETTINGS.git.use_explicit_merge_commit = True
     yield
-    config.SETTINGS.git.use_explicit_merge_commit = False
+    config.SETTINGS.git.use_explicit_merge_commit = initial_use_explicit_merge_commit
